Some fixes for single precsision

+++ matRad/plotting/matRad_computeVoiContours.m
@@ -38,17 +38,17 @@ for ctScen = 1:ct.numOfCtScen
         mask(cst{s,4}{ctScen}) = 1;
         for slice = 1:ct.cubeDim(1)
             if any(any(mask(slice,:,:) > 0))
-                cst{s,7}{1,ctScen}{slice,1} = contourc(squeeze(mask(slice,:,:)),.5*[1 1]);
+                cst{s,7}{1,ctScen}{slice,1} = contourc(double(squeeze(mask(slice,:,:))),.5*[1 1]);
             end
         end
         for slice = 1:ct.cubeDim(2)
             if any(any(mask(:,slice,:) > 0))
-                cst{s,7}{1,ctScen}{slice,2} = contourc(squeeze(mask(:,slice,:)),.5*[1 1]);
+                cst{s,7}{1,ctScen}{slice,2} = contourc(double(squeeze(mask(:,slice,:))),.5*[1 1]);
             end
         end
         for slice = 1:ct.cubeDim(3)
             if any(any(mask(:,:,slice) > 0))
-                cst{s,7}{1,ctScen}{slice,3} = contourc(squeeze(mask(:,:,slice)),.5*[1 1]);
+                cst{s,7}{1,ctScen}{slice,3} = contourc(double(squeeze(mask(:,:,slice))),.5*[1 1]);
             end
         end
         if matRad_cfg.logLevel > 2
